requirements-runtime.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=runtime --output-file=requirements-runtime.txt --strip-extras pyproject.toml
#
--index-url https://gitea.woggioni.net/api/packages/woggioni/pypi/simple/
--extra-index-url https://pypi.org/simple

asgiref==3.8.1
    # via
    #   django
    #   opentelemetry-instrumentation-asgi
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via granian
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
django==5.1.1
    # via
    #   django_otel_poc (pyproject.toml)
    #   djangorestframework
djangorestframework==3.15.2
    # via django_otel_poc (pyproject.toml)
googleapis-common-protos==1.65.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
granian==1.6.0
    # via django_otel_poc (pyproject.toml)
grpcio==1.66.1
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.10
    # via requests
importlib-metadata==8.4.0
    # via opentelemetry-api
opentelemetry-api==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-system-metrics
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.27.0
    # via django_otel_poc (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.27.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.48b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-system-metrics
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-asgi==0.48b0
    # via django_otel_poc (pyproject.toml)
opentelemetry-instrumentation-django==0.48b0
    # via django_otel_poc (pyproject.toml)
opentelemetry-instrumentation-logging==0.48b0
    # via django_otel_poc (pyproject.toml)
opentelemetry-instrumentation-system-metrics==0.48b0
    # via django_otel_poc (pyproject.toml)
opentelemetry-instrumentation-wsgi==0.48b0
    # via opentelemetry-instrumentation-django
opentelemetry-proto==1.27.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.27.0
    # via
    #   django_otel_poc (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.48b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.48b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-wsgi
pillow==10.4.0
    # via django_otel_poc (pyproject.toml)
protobuf==4.25.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psutil==6.0.0
    # via opentelemetry-instrumentation-system-metrics
psycopg2-binary==2.9.9
    # via django_otel_poc (pyproject.toml)
python-json-logger==2.0.7
    # via django_otel_poc (pyproject.toml)
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via opentelemetry-sdk
urllib3==2.2.3
    # via requests
uvloop==0.20.0
    # via granian
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
